Split off compress from files.py
Stream#delete_stream() -> Stream#delete()
